[ALPS07247746] modify kernel time accuracy of driver dump
1. modify kernel time accuracy of driver dump
2. debug sop tag name

MTK-Commit-Id: 75e2441e13a5835ff52d9695d5f99af70a3e28f3

CR-Id: ALPS07247746
Feature: BT Driver

+++ include/btmtk_define.h
@@ -377,8 +377,8 @@ struct bt_utc_struct {
 	struct rtc_time tm;
 	u32 usec;
 	u32 sec;
-	u32 ksec;
-	u32 knsec;
+	unsigned long ksec;
+	unsigned long knsec;
 };
 
 #define BT_DOWNLOAD	1
